+++ pcs/lib/reports.py
@@ -2461,3 +2461,11 @@ def fencing_level_does_not_exist(level, target_type, target_value, devices):
             "devices": devices,
         }
     )
+
+def ambiguous_host_specification(host_list):
+    return ReportItem.error(
+        report_codes.AMBIGUOUS_HOST_SPECIFICATION,
+        info={
+            "host_list": host_list,
+        }
+    )